Move ParseError to shared folder
