+++ src/shared/ParseError.ts
@@ -1,0 +1,11 @@
+export class ParseError extends Error {
+  private constructor(error: Error) {
+    super(error.message)
+
+    Object.setPrototypeOf(this, ParseError.prototype)
+  }
+
+  public static fromError(error: Error) {
+    return new this(error)
+  }
+}
