+++ src/shared/http/HttpResult.ts
@@ -1,5 +1,5 @@
 import { AsyncResult } from '@shared/result'
 import { HttpError } from '@shared/http/HttpError'
-import { ParseError } from './ParseError'
+import { ParseError } from '../ParseError'
 
 export type HttpResult<T> = AsyncResult<T, HttpError | ParseError>
